Removed command error chat notification
This just gives false positives for RegisterCommands that uses +commandName.

Requesting removal of this pice of code, since i dont se it bringing anything good to the table to be honest.

Not sure if this is intended code, and what reason there is behind keeping it there?

In my perspective it's just bloat.